+++ inst/scripts/make-exampleTSSr.R
@@ -48,7 +48,7 @@ exampleTSSr@PromoterShift <- list()
 
 exampleTSSr <- mergeSamples(exampleTSSr)
 exampleTSSr <- normalizeTSS(exampleTSSr)
-exampleTSSr <- filterTSS(exampleTSSr, method = "TPM", tpmLow = 0.1)
+exampleTSSr <- filterTSS(exampleTSSr, method = "TPM", tpmLow = 2)
 exampleTSSr <- clusterTSS(
     exampleTSSr,
     method = "peakclu",